requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
PyYAML==6.0
scikit_learn==1.2.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
